Game Manager run, clock fix to stop 'catch up' effect that is noticed when dragging game window and 'freezing' the display.

+++ GameEngine/EventStep.cpp
@@ -12,17 +12,17 @@ namespace df {
 		m_step_count = 0;
 	}
 
-	EventStep::EventStep(int init_step_count) {
+	EventStep::EventStep(long long init_step_count) {
 
 		setType(STEP_EVENT);
 		m_step_count = init_step_count;
 	}
 
-	void EventStep::setStepCount(int new_step_count) {
+	void EventStep::setStepCount(long long new_step_count) {
 		m_step_count = new_step_count;
 	}
 
-	int EventStep::getStepCount() const {
+	long long EventStep::getStepCount() const {
 		return m_step_count;
 	}
